Change the encoding of non-ascii Uptane metadata
The JSON serialization spec is not canonical. For example the string "€" can be
encoded as either "€" or "\u20ac". Given that JSON encodes a data structure
into a list of Unicode codepoints, we then encode that into the sequence of
bytes that get signed using UTF-8. For the two examples above, these bytes are:

34 226 130 172 34          " <3 byte U+20ac> "
34 92 117 50 48 97 99 34   " \ u 2 0 a c "

It turns out that libtuf was using the former encoding, and Aktualizr was using
the latter.

This change sets the emitUTF8 flag on the Json library to get the other
behaviour. The exact change in encoding rules can be found in
valueToQuotedStringN() in jsoncpp.cc.

Risks: This changes Aktualizr's metadata validation behaviour, so previously
valid metadata will be rejected. However, I don't believe that there is a
backend in production that generates metadata which is:
 - Valid according to the old behaviour
 - Invalid according to the new behaviour

uptane_generator _does_, but I don't know of a production deployment using it.

Note that scanning existing Uptane metadata for bytes with the high bit set is
not sufficient to demonstrate safety: it is necessary to parse it, then check for
any codepoints above U+7F in strings of the resulting structure.