If installing on the primary fails, skip secondaries
If installation on the primary ECU fails, then don't attempt to install
on the secondaries. Note that this only takes effect if the failure
happens before the primary is rebooted to apply the update. In the case
of an OSTree update this is likely to be OK, since most of the work that
can fail (e.g. network downloads) happens before the reboot.