Contain CommandRunner children with a process group
This fixes an issue where Aktualizr can't abort docker pull

If an (online) Docker installation is in process when an offline update
becomes available, then Aktualizr needs to abort the on-going
docker-compose pull operation. The code to do this didn't work because
/usr/bin/docker-compose is a wrapper that doesn't propagate the
terminate signal to its children. Fix that by running the whole thing in
a process group.