Load TLS certs earlier in provisioning process
The main provisioning process happens in SotaUptaneClient::initialize()
after pending updates have been installed. After installing an update,
aktualizr reports the manifest to the server, but this was failing with
the following error because the TLS credentials were not setup:

  curl error 60 (http code 0): SSL peer certificate or SSH remote key was not OK

Fix this loading the TLS credentials in the 'Prepare' phase if they are
present.