Move CI from old Ubuntu images to Debian
Aktualizr CI builds are running on a very old version of Ubuntu. Bionic
was released in April 2018 and ended support May 2023. Xenial was
released in April 2016 and while it does have 10 years of support until
2026 it is stil ancient.

Instead use Debian (old)stable for the main test image. At the moment
we don't build cleanly on Debian Stable (Bookworm). I've got a set of
changes locally to fix that, but the first job is to get the CI clean,
then we can look at moving it forward.

Finally, this nukes the old codecov stuff. No-one ever looked at those
results, and if we want something that is low-cost to maintain we need
to simplify where possible.